--- app/Livewire/App/Settings/Modals/ColorModal.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Livewire\App\Settings\Modals;
-
-use Livewire\Component;
-
-class ColorModal extends Component
-{
-    public function render()
-    {
-        return view('livewire.app.settings.modals.color-modal');
-    }
-}
