+++ resources/views/livewire/app/settings/links.blade.php
@@ -1,0 +1,20 @@
+<?php
+
+use Livewire\Volt\Component;
+
+
+new class extends Component {
+
+
+
+}
+
+?>
+
+<div>
+
+    soon
+
+
+
+</div>
